+++ kleiner-brauhelfer/mainwindow.ui
@@ -60,7 +60,7 @@
         <item>
          <widget class="QTabWidget" name="tabMain">
           <property name="currentIndex">
-           <number>7</number>
+           <number>0</number>
           </property>
           <widget class="TabSudAuswahl" name="tabSudAuswahl">
            <attribute name="icon">
@@ -661,6 +661,9 @@
    <property name="text">
     <string>Rohstoffe</string>
    </property>
+   <property name="whatsThis">
+    <string>actionrohstoffe</string>
+   </property>
    <property name="visible">
     <bool>true</bool>
    </property>
@@ -678,6 +681,9 @@
    </property>
    <property name="text">
     <string>Brauübersicht</string>
+   </property>
+   <property name="whatsThis">
+    <string>actionbrauuebersicht</string>
    </property>
    <property name="visible">
     <bool>true</bool>
@@ -697,6 +703,9 @@
    </property>
    <property name="text">
     <string>Ausrüstung</string>
+   </property>
+   <property name="whatsThis">
+    <string>actionausruestung</string>
    </property>
    <property name="visible">
     <bool>true</bool>
